% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as.ffdf.ffdf}
\alias{as.ffdf.ffdf}
\title{Trivial implementation, but very handy}
\usage{
\method{as.ffdf}{ffdf}(x, ...)
}
\arguments{
\item{x}{ffdf object}

\item{...}{not used.}
}
\description{
Coerce a ffdf object to an ffdf object.
}
