\name{as.ram.ffdf}
\alias{as.ram.ffdf}
\title{As ram for an ffdf to explicitely get your ffdf as a data frame}
\usage{
  \method{as.ram}{ffdf} (x, ...)
}
\arguments{
  \item{x}{an object of class ffdf}

  \item{...}{not used.}
}
\value{
  a data.frame in RAM
}
\description{
  Coerce a ffdf object to an data.frame object.
}

