\name{c.ff}
\alias{c.ff}
\title{Concatenate \code{ff} vectors}
\usage{
\method{c}{ff}(...)
}
\arguments{
  \item{...}{\code{ff} ff vectors to be concatenated}
}
\value{
a new \code{ff} object, data is physically copied
}
\description{
Concatenate \code{ff} vectors
}
\seealso{
\code{\link{ffappend}}
}

