% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compact.R
\name{compact}
\alias{compact}
\alias{compact.ff}
\alias{compact.ffdf}
\title{Compact a ff vector or ffdf data frame}
\usage{
\method{compact}{ff}(x, use.na = TRUE, ...)
}
\arguments{
\item{x}{\code{ff} or \code{ffdf} object}

\item{use.na}{\code{logical} if TRUE the resulting ff vector can contain NA, otherwise this is not checked}

\item{...}{other parameters}
}
\value{
compact cloned ff vector, or original if no compacting can be done
}
\description{
Compact takes a ff vector and tries to use the smallest binary data type for this vector.
}
