% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_ff.R
\name{cut.ff}
\alias{cut.ff}
\title{Convert Numeric ff vector to factor ff}
\usage{
\method{cut}{ff}(x, breaks, ...)
}
\arguments{
\item{x}{a (numeric) ff object that will be cut into pieces}

\item{breaks}{specifies the breaks for cutting this}

\item{...}{other parameters that can be given to \code{\link{cut.default}}}
}
\value{
ff a new \code{\link{ff}} object with the newly created factor
}
\description{
cut divides the range of x into intervals and codes the values in x according to which interval they fall. The leftmost interval corresponds to level one, the next leftmost to level two and so on.
}
\details{
The \code{cut} method for ff with the behaviour of \code{link{cut}}
}
\seealso{
cut
}

