% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ops.R
\name{+.ff_vector}
\alias{*.ff_vector}
\alias{+.ff_vector}
\alias{-.ff_vector}
\alias{/.ff_vector}
\alias{Arithmetic}
\alias{\%/\%.ff_vector}
\alias{\%\%.ff_vector}
\alias{^.ff_vector}
\title{Arithmetic Operators for ff vectors}
\usage{
\method{+}{ff_vector} (x, y)

\method{-}{ff_vector} (x, y)

\method{*}{ff_vector} (x, y)

\method{/}{ff_vector} (x, y)

\method{^}{ff_vector} (x, y)

\method{\%\%}{ff_vector} (x, y)

\method{\%/\%}{ff_vector} (x, y)
}
\arguments{
\item{x}{either a numeric \code{ff_vector} or a vector of length 1 in RAM in which case y should be an \code{ff_vector}}

\item{y}{either a numeric \code{ff_vector} or a vector of length 1 in RAM in which case x should be an \code{ff_vector}}
}
\value{
an \code{ff_vector}. For the definition of the operators see the base package of R.
}
\description{
These binary operators perform arithmetic on numeric ff vectors. 
Arith family: \cr
\itemize{
\item{Arith: }{\code{"+", "-", "*", "/", "^", "\%\%", "\%/\%"}}
}
The operators require either \code{x} or \code{y} to be an \code{ff_vector} or both. In case either \code{x} or \code{y} is not an \code{ff_vector},
the other object needs to be of length 1. Recycling is not implemented.
}

