% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ops.R
\name{abs.ff_vector}
\alias{abs.ff_vector}
\alias{Math}
\alias{sign.ff_vector}
\alias{sqrt.ff_vector}
\alias{ceiling.ff_vector}
\alias{floor.ff_vector}
\alias{trunc.ff_vector}
\alias{log.ff_vector}
\alias{log10.ff_vector}
\alias{log2.ff_vector}
\alias{log1p.ff_vector}
\alias{acos.ff_vector}
\alias{acosh.ff_vector}
\alias{asin.ff_vector}
\alias{asinh.ff_vector}
\alias{atan.ff_vector}
\alias{atanh.ff_vector}
\alias{exp.ff_vector}
\alias{expm1.ff_vector}
\alias{cos.ff_vector}
\alias{cosh.ff_vector}
\alias{sin.ff_vector}
\alias{sinh.ff_vector}
\alias{tan.ff_vector}
\alias{tanh.ff_vector}
\alias{gamma.ff_vector}
\alias{lgamma.ff_vector}
\alias{digamma.ff_vector}
\alias{trigamma.ff_vector}
\alias{round.ff_vector}
\alias{signif.ff_vector}
\title{Math for ff vectors}
\usage{
\method{abs}{ff_vector} (x)

\method{sign}{ff_vector} (x)

\method{sqrt}{ff_vector} (x)

\method{ceiling}{ff_vector} (x)

\method{floor}{ff_vector} (x)

\method{trunc}{ff_vector} (x, ...)

\method{log10}{ff_vector} (x)

\method{log2}{ff_vector} (x)

\method{log1p}{ff_vector} (x)

\method{acos}{ff_vector} (x)

\method{acosh}{ff_vector} (x)

\method{asin}{ff_vector} (x)

\method{asinh}{ff_vector} (x)

\method{atan}{ff_vector} (x)

\method{atanh}{ff_vector} (x)

\method{exp}{ff_vector} (x)

\method{expm1}{ff_vector} (x)

\method{cos}{ff_vector} (x)

\method{cosh}{ff_vector} (x)

\method{sin}{ff_vector} (x)

\method{sinh}{ff_vector} (x)

\method{tan}{ff_vector} (x)

\method{tanh}{ff_vector} (x)

\method{gamma}{ff_vector} (x)

\method{lgamma}{ff_vector} (x)

\method{digamma}{ff_vector} (x)

\method{trigamma}{ff_vector} (x)

\method{log}{ff_vector} (x, base)

\method{round}{ff_vector} (x, digits)

\method{signif}{ff_vector} (x, digits)
}
\arguments{
\item{x}{a numeric \code{ff_vector}}

\item{...}{for \code{trunc}, currently not used}

\item{base}{base for \code{log}}

\item{digits}{digits for \code{round} and \code{signif}}
}
\value{
an \code{ff_vector}. For the definition of the operators see the base package of R.
}
\description{
These mathematical functions implement \code{ff_vector} specific math and handle the following functions from the
Math family: \cr
\itemize{
\item{Math: }{\code{"abs", "sign", "sqrt", "ceiling", "floor", "trunc", "log", "log10", "log2", "log1p", "acos", "acosh", "asin", "asinh", "atan", "atanh", "exp", "expm1", "cos", "cosh", "sin", "sinh", "tan", "tanh", "gamma", "lgamma", "digamma", "trigamma"}}
\item{Math2: }{\code{"round", "signif"}}
}
The operators require \code{x} to be an \code{ff_vector}.
}
