% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist2d.R
\name{hist2d.ff}
\alias{hist2d.ff}
\title{compute and plot 2 dimensional histogram of ff data}
\usage{
hist2d.ff(x, y = NULL, nbins = 100, show = TRUE)
}
\description{
function interface modeled after gplots::hist2d
}
\keyword{internal}

