% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laf_to_ffdf.R
\name{laf_to_ffdf}
\alias{laf_to_ffdf}
\title{Use LaF to import data into \code{ffdf} data.frame}
\usage{
laf_to_ffdf(laf, x = NULL, nrows = 1e+05, transFUN = NULL, ...)
}
\arguments{
\item{laf}{laf object pointing to a csv or fwf file}

\item{x}{optional, \code{\link{ffdf}} object where laf data should be appended to.}

\item{nrows, }{number of rows per block, passed on to \code{next_block}}

\item{transFUN}{NULL or a function that is called on each data.frame chunk which is read in using \code{next_block}. 
This can be used for filtering and data transformations.}

\item{...}{passed on to \code{next_block}}
}
\description{
Use \code{LaF} to import data into a \code{\link{ffdf}} data.frame
}

