\name{risplit}
\alias{risplit}
\alias{risplit.default}
\alias{risplit.ff}
\alias{risplit.ffdf}
\title{split using ri}
\usage{
  risplit(x, f)

  risplit.default(x, f)

  \method{risplit}{ff} (x, f)

  \method{risplit}{ffdf} (x, f)
}
\arguments{
  \item{f}{\code{ff factor} object used for splitting}
}
\description{
  split using ri
}
\keyword{internal}

