% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as.ram.ffdf}
\alias{as.ram.ffdf}
\title{As ram for an ffdf to get your ffdf as a data frame in RAM}
\usage{
\method{as.ram}{ffdf}(x, ...)
}
\arguments{
\item{x}{an object of class ffdf}

\item{...}{not used.}
}
\value{
a data.frame in RAM
}
\description{
Load your ffdf object in RAM into a data.frame.
}
