\name{chunkexpr}
\alias{chunkexpr}
\title{Chunk an expression to be used in a chunked for loop}
\usage{
  chunkexpr(x, expr, i = ".i", prefix = "")
}
\arguments{
  \item{x}{\code{character} with vars}

  \item{expr}{\code{expression} vector}

  \item{i}{name of index}

  \item{prefix}{prefix for variables to be replaced.}
}
\description{
  Chunk an expression to be used in a chunked for loop
}
\keyword{internal}

