\name{table.ff}
\alias{table.ff}
\title{table...}
\usage{table.ff(..., exclude=if (useNA == "no") c(NA, NaN), useNA=c("no", "ifany",
    "always"), dnn=list.names(...), deparse.level=1)
}
\description{table.ff uses the cross-classifying factors to build a contingency table of the 
counts at each combination of factor levels.}
\details{Details}
\seealso{\code{\link{table}}}
\value{\code{table} object}
\arguments{\item{...}{\code{ff} factors}
\item{exclude}{see \code{\link{table}}}
\item{useNA}{see \code{\link{table}}}
\item{dnn}{see \code{\link{table}}}
\item{deparse.level}{see \code{\link{table}}}
}

