\name{unique.ff}
\alias{unique.ff}
\title{Unique values...}
\usage{\method{unique}{ff}(x, incomparables=FALSE, fromLast=FALSE, ...)
}
\description{Unique values}
\value{vector with unique values of \code{x}}
\arguments{\item{x}{\code{ff} object}
\item{incomparables}{?}
\item{fromLast}{?}
\item{...}{Further arguments passed to \code{\link{unique}}}
}

